refactor(agents): split before-tool-call state
