+++ src/agents/agent-tool-definition-adapter.after-tool-call.fires-once.test.ts
@@ -101,6 +101,7 @@ async function loadFreshAfterToolCallModulesForTest() {
   vi.doMock("./agent-tools.before-tool-call.state.js", () => ({
     carryLoopWarningToError: (error: unknown) => error,
     carryLoopWarningToToolResult: (result: unknown) => result,
+    resolveLoopWarningError: (error: unknown) => error,
     consumeAdjustedParamsForToolCall: beforeToolCallMocks.consumeAdjustedParamsForToolCall,
     consumeLoopWarningForToolCall: vi.fn(() => undefined),
     consumePreExecutionBlockedToolCall: vi.fn(() => false),
@@ -108,6 +109,7 @@ async function loadFreshAfterToolCallModulesForTest() {
     consumeStructuredReplaySafeToolCall: vi.fn(() => false),
     peekAdjustedParamsForToolCall: vi.fn(() => undefined),
     peekPreExecutionBlockedToolCall: vi.fn(() => false),
+    recordPreExecutionBlockedToolCall: vi.fn(),
   }));
   vi.doMock("./agent-tools.before-tool-call.js", () => ({
     BeforeToolCallBlockedError: beforeToolCallMocks.BeforeToolCallBlockedError,
